fix: 내가 쓴 글 조회에서 댓글수가 제대로 조회 안되는 버그 수정 (#426)
